Sonarcloud and build warnings
This commit finalises my initial work for src/CryptoNote until https://sonarcloud.io/summary/overall?branch=branch-sonarcloud&id=bstera_conceal-core is refreshed. Further work will come in later pull requests to give us time to test these changes

+++ include/ITransaction.h
@@ -109,7 +109,7 @@ public:
   virtual size_t addInput(const AccountKeys& senderKeys, const transaction_types::InputKeyInfo& info, KeyPair& ephKeys) = 0;
 
   virtual size_t addOutput(uint64_t amount, const AccountPublicAddress& to) = 0;
-  virtual size_t addOutput(uint64_t amount, const std::vector<AccountPublicAddress>& to, uint32_t requiredSignatures, uint32_t term = 0) = 0;
+  virtual size_t addOutput(uint64_t amount, const std::vector<AccountPublicAddress>& to, uint8_t requiredSignatures, uint32_t term = 0) = 0;
   virtual size_t addOutput(uint64_t amount, const KeyOutput& out) = 0;
   virtual size_t addOutput(uint64_t amount, const MultisignatureOutput& out) = 0;
